EndianSwap

+++ RCore/src/RMemory/RMemory.cpp
@@ -1,7 +1,7 @@
 #include "RMemory.h"
 #include <assert.h>
 
-UINT64 RMemory::align(UINT64 value, UINT64 alignment)
+size_t RMemory::align(size_t value, size_t alignment)
 {
 	if ((alignment == 0) || (alignment & (alignment - 1)) != 0)
 		throw("Alignment is not a power of 2");
